+++ scripts/JSRootCore.js
@@ -122,24 +122,6 @@
          kIsAverage     : JSROOT.BIT(18)  // Bin contents are average (used by Add)
    };
 
-   JSROOT.TObjectBits = {
-         kCanDelete     : JSROOT.BIT(0),   // if object in a list can be deleted
-         kMustCleanup   : JSROOT.BIT(3),   // if object destructor must call RecursiveRemove()
-         kObjInCanvas   : JSROOT.BIT(3),   // for backward compatibility only, use kMustCleanup
-         kIsReferenced  : JSROOT.BIT(4),   // if object is referenced by a TRef or TRefArray
-         kHasUUID       : JSROOT.BIT(5),   // if object has a TUUID (its fUniqueID=UUIDNumber)
-         kCannotPick    : JSROOT.BIT(6),   // if object in a pad cannot be picked
-         kNoContextMenu : JSROOT.BIT(8),   // if object does not want context menu
-         kInvalidObject : JSROOT.BIT(13),  // if object ctor succeeded but object should not be used
-         kSingleKey     : JSROOT.BIT(0),   // write collection with single key
-         kOverwrite     : JSROOT.BIT(1),   // overwrite existing object with same name
-         kWriteDelete   : JSROOT.BIT(2),   // write object, then delete previous key with same name
-         kIsOnHeap      : 0x01000000,      // object is on heap
-         kNotDeleted    : 0x02000000,      // object has not been deleted
-         kZombie        : 0x04000000,      // object ctor failed
-         kBitMask       : 0x00ffffff,
-   };
-
    JSROOT.EAxisBits = {
          kTickPlus      : JSROOT.BIT(9),
          kTickMinus     : JSROOT.BIT(10),
@@ -1014,7 +996,7 @@
       // check object type and add methods if needed
       if (('fBits' in obj) && !('TestBit' in obj)) {
          obj['TestBit'] = function (f) { return (this['fBits'] & f) != 0; };
-         obj['InvertBit'] = function (f) { this['fBits'] = this['fBits'] ^ (f & JSROOT.TObjectBits.kBitMask); };
+         obj['InvertBit'] = function (f) { this['fBits'] = this['fBits'] ^ (f & 0xffffff); };
       }
 
       if (!obj_typename) {
@@ -1255,7 +1237,7 @@
             // Note that if h1 has Sumw2 set, Sumw2 is automatically called for this
             // if not already set.
             if (!h1 || typeof(h1) == 'undefined') {
-               alert("Add : Attempt to add a non-existing histogram");
+               alert("Add : Attempt to add( a non-existing histogram");
                return false;
             }
             if (!c1 || typeof(c1) == 'undefined') c1 = 1;
@@ -2070,17 +2052,19 @@
 
    JSROOT.log10 = function(n) {
       return Math.log(n) / Math.log(10);
-   };
+   }
 
    // it is important to run this function at the end when all other
    // functions are available
-   (function() {
+   JSROOT.Init = function() {
       function window_on_load(func) {
-         if (func==null) return;
-         if (document.attachEvent ? document.readyState === 'complete' : document.readyState !== 'loading')
-            func();
-         else
-            window.onload = func;
+         if (func!=null) {
+            if (document.attachEvent ? document.readyState === 'complete' : document.readyState !== 'loading')
+               func();
+            else
+               window.onload = func;
+         }
+         return JSROOT;
       }
 
       var scripts = document.getElementsByTagName('script');
@@ -2125,11 +2109,12 @@
               }
             });
 
-         return;
+         return this;
       }
-   })();
+      return this;
+   }
 
-   return JSROOT;
+   return JSROOT.Init();
 
 }));